Implementing a third level. Level 1 is the moving circle, level 2 is the original attempt at making a pole-star style game, and level 3 is the latest road code.
